3.3.4.2 ongoing

+++ src/AlpacaServer.cpp
@@ -336,7 +336,7 @@ bool AlpacaServer::GetParam(AsyncWebServerRequest *request, const char *name, do
     int32_t index = _paramIndex(request, name, spelling);
     if (index >= 0)
     {
-        return sscanf(request->arg(static_cast<int32_t>(index)).c_str(), "%lf", &value) == 1;
+        return sscanf(request->arg(static_cast<int>(index)).c_str(), "%lf", &value) == 1;
     }
     return false;
 }
@@ -347,7 +347,7 @@ bool AlpacaServer::GetParam(AsyncWebServerRequest *request, const char *name, fl
     int32_t index = _paramIndex(request, name, spelling);
     if (index >= 0)
     {
-        return sscanf(request->arg(static_cast<int32_t>(index)).c_str(), "%f", &value) == 1;
+        return sscanf(request->arg(static_cast<int>(index)).c_str(), "%f", &value) == 1;
     }
     return false;
 }
@@ -359,7 +359,7 @@ bool AlpacaServer::GetParam(AsyncWebServerRequest *request, const char *name, in
     int32_t index = _paramIndex(request, name, spelling);
     if (index >= 0)
     {
-        return sscanf(request->arg(index).c_str(), "%i", &value) == 1;
+        return sscanf(request->arg(static_cast<int>(index)).c_str(), "%i", &value) == 1;
     }
     return false;
 }
@@ -370,7 +370,7 @@ bool AlpacaServer::GetParam(AsyncWebServerRequest *request, const char *name, ui
     int32_t index = _paramIndex(request, name, spelling);
     if (index >= 0)
     {
-        return sscanf(request->arg(index).c_str(), "%u", &value) == 1;
+        return sscanf(request->arg(static_cast<int>(index)).c_str(), "%u", &value) == 1;
     }
     return false;
 }
@@ -381,7 +381,7 @@ bool AlpacaServer::GetParam(AsyncWebServerRequest *request, const char *name, ch
     int32_t index = _paramIndex(request, name, spelling);
     if (index >= 0)
     {
-        request->arg(index).toCharArray(buffer, buffer_size);
+        request->arg(static_cast<int>(index)).toCharArray(buffer, buffer_size);
         buffer[buffer_size - 1] = '\0';
         return true;
     }
